refactor: extract `Core` trait from `BaseNode`, delete `Node` trait
Separates local skip-graph algorithms (new `Core` trait + `BaseCore`
impl owning id, mem_vec, and lookup table) from network orchestration
(`BaseNode` now composes `Box<dyn Core>`). Splits tests: `core_test.rs`
for local-algorithm tests against `BaseCore` (no mock network needed),
`search_by_id_test.rs` for the two `process_incoming_event` integration
cases. Renames `get_identifier`/`get_membership_vector` to `id`/`mem_vec`.
